Weather Application Using API
This is A Weather Application that can be used to send emails to employees of the current weather on a particluar day with used on openweather API. This Applicatiion can also indicate to these employees if they should go to work or not.
I hope this application is useful to you!